+++ pkg/filen/analyzer_test.go
